+++ core/domain/src/main/java/es/joshluq/kmsafe/domain/model/OdometerRecord.kt
@@ -10,7 +10,7 @@ data class OdometerRecord(
     val id: String,
     val contractId: String = "",
     val timestamp: Long,
-    val odometerValue: Int,
+    val odometerValue: Double,
     val isInitialRecord: Boolean,
     val label: String? = null,
     val fuelAmount: Double? = null,
